pyproject.toml
[tool.poetry]
name = "image_classification"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.0.1"
torch = "^1.9.0"
torchvision = "^0.10.0"
pydantic = "^1.8.2"
pytorch-lightning = ["^1.4.5", "!=1.4.6", "!=1.4.7"]
onnx = "^1.10.1"
Bottleneck = "^1.3.2"
"ruamel.yaml" = "^0.17.13"
loguru = "^0.5.3"
torchmetrics = "^0.5.1"
albumentations = "^1.0.3"
tensorboard = "^2.6.0"
torch-tb-profiler = "^0.2.1"
adabelief-pytorch = "^0.2.1"
hydra-core = "^1.1.1"
hydra-optuna-sweeper = "^1.1.1"
scikit-learn = "^0.24.2"
matplotlib = "^3.4.3"
fastcore = "^1.3.26"
lightning-bolts = "^0.4.0"
neptune-client = "^0.14.3"

[tool.poetry.dev-dependencies]
pre-commit = "^2.15.0"
pytest = "^6.2.5"

[tool.black]
line-length = 80

[flake8]
max-line-length = 80
ignore = "E203,E266,E501,W503,F403,F401" # For black compatibility
max-complexity = 18

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
train = "image_classification.trainer:train"
